Release v4.0.0 — engine v4 is the default; includes v3.7.0/3.8.0 hardening
Engine deep analysis (docs/reports/report-engine-deep-analysis-2026-08-24.md, ENG-01..17)
and the clean-room engine v4 program (docs/reports/report-engine-v4-design-2026-08-25.md,
handoff docs/reports/report-engine-v4-progress.md, waves W1-W9, ISS-2025-0423..0490):

- v3.7.0/3.8.0: repeat/length/float typing fixes, tail-iterative term walkers (1M-element
  lists), bounded-memory machine, ResourceGuard budget inside meta-calls, first-argument
  indexing, per-engine flags.
- v4 (core.engine.v4, default since 4.0.0; v2 MachineSolver selectable one release via
  -Djprolog.engine=v2 / -Pengine-v2): cell-based variables, clause skeletons, cycle-safe
  unification (rational trees), generations clause store, native meta-calls/DCG/generators,
  Prolog prelude (apply/lists/pairs/coroutining), yall lambdas, attributed-variable wake queue
  (freeze/when/dif/CLP(FD)), linear tabling with completion, modules (system/user/library),
  per-engine streams/operators/spy/profiler, ISO writer and answer printing with residual
  goals, threads on v4 machines, tracer without disabled fast paths, batch console.
- Recursive QuerySolver, CutStatus, LayeredMap and the legacy control-construct built-ins
  deleted; BuiltInWithContext reduced to the adapter over SolverContext/EngineContext.
- Reference Manual (docs/guides/guide-builtin-manual.md/.pdf, generated by
  tools/build-manual.sh) documenting every default operator and built-in with verified examples.

Verified: mvn test 1214/1214 on both engine legs, test_all_examples.sh 20/20 on both.

+++ src/main/java/it/denzosoft/jprolog/builtin/arithmetic/Succ.java
@@ -92,7 +92,7 @@ public class Succ implements BuiltIn {
             // END_CHANGE: ISS-2025-0190
 
             Map<String, Term> newBindings = new HashMap<>(bindings);
-            if (query.getArguments().get(0).unify(new it.denzosoft.jprolog.core.terms.Number((double) int1Value), newBindings)) {
+            if (query.getArguments().get(0).unify(new it.denzosoft.jprolog.core.terms.Number((long) int1Value)   /* ISS-2025-0424 */, newBindings)) {
                 solutions.add(newBindings);
                 return true;
             } else {
@@ -115,7 +115,7 @@ public class Succ implements BuiltIn {
             // END_CHANGE: ISS-2025-0190
             
             Map<String, Term> newBindings = new HashMap<>(bindings);
-            if (query.getArguments().get(1).unify(new it.denzosoft.jprolog.core.terms.Number((double) int2Value), newBindings)) {
+            if (query.getArguments().get(1).unify(new it.denzosoft.jprolog.core.terms.Number((long) int2Value)   /* ISS-2025-0424 */, newBindings)) {
                 solutions.add(newBindings);
                 return true;
             } else {
